requirements.txt
boto3==1.38.8
botocore==1.38.8
Flask==3.1.1
flask_cors==5.0.1
flask_sock==0.7.0
Pillow==9.4.0
pymongo==4.13.0
python-dotenv==1.1.0
Requests==2.32.3
